requirements.txt
absl-py==1.2.0
Adafruit-Blinka==8.4.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-register==1.9.14
adafruit-circuitpython-requests==1.12.8
adafruit-circuitpython-typing==1.8.1
Adafruit-PlatformDetect==3.29.0
Adafruit-PureIO==1.1.9
addict==2.4.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
commonmark==0.9.1
cycler==0.11.0
Cython==0.29.32
fonttools==4.37.1
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.3
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
oauthlib==3.2.0
onnx==1.12.0
onnxsim==0.4.8
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyftdi==0.54.0
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
pyusb==1.2.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.5.1
rplidar-roboticia==0.9.5
rsa==4.9
scipy==1.9.1
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
